# Timeout if populating iterators takes too long