Right now promql won't time out a request if populating the iterators
takes a long time.

Fixes #4289