@juliusv 2018-06-19
Hehe, my https://github.com/prometheus/prometheus/pull/4290 was a minute earlier :P

But I'll retract it.

:+1: thanks!