@juliusv 2018-06-19
/cc @brian-brazil since he had plans on doing something like this already