@brian-brazil 2018-06-19
No, this was not on my todo list. If you're spending a long time here, something is wrong with local storage.