@juliusv 2018-06-19
@brian-brazil The query was `{__name__=~".+"}[5d]`, so it selects all series. I could see how that could take a long time when you have millions of series.