@brian-brazil 2018-06-20
I wonder what the performance cost is of checking this on every single series. What do the benchmarks say?