@jacksontj 2018-06-20
Benchmark code: https://play.golang.org/p/S-nPo-Emx3X

results:

```
[jacksontj@localhost selectBench]$ go test -run=x -bench=.
goos: linux
goarch: amd64
pkg: github.com/jacksontj/test/selectBench
BenchmarkSelect-8   	200000000	         7.93 ns/op
PASS
ok  	github.com/jacksontj/test/selectBench	2.388s
```

So about 8ns per check.