@brian-brazil 2018-06-20
What does that do to the existing promql benchmarks?