@jacksontj 2018-06-20
Doing some quick math, in the case of 10MM timeseries this would be:
```
8ns * 10000000 = 80ms
``

I ran the promql benchmarks, but they don't seem to yield stable results (meaning I run it twice on the same code before my change and the values seem to vary quite a bit). Regardless, here is the comparison:

```
tjackson@tjackson-ThinkPad-X1-Carbon-5th:~/workspace/golang/src/github.com/prometheus/prometheus/promql$ benchcmp -changed -mag old new
benchmark                                                                                                             old ns/op       new ns/op       delta
BenchmarkRangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=100-4                                6234048         10906837        +74.96%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-4                                                                    345191          214007          -38.00%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-4                                                                 2203144737      1474983237      -33.05%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-4                                                                         2312106         1564468         -32.34%
BenchmarkRangeQuery/expr=-a_hundred,steps=10-4                                                                        2247544         1570975         -30.10%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-4                                                   179523          245054          +36.50%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-4                                                                    1132444         1527657         +34.90%
BenchmarkRangeQuery/expr=rate(a_ten[1m])_+_rate(b_ten[1m]),steps=1-4                                                  542039          405232          -25.24%
BenchmarkRangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=10-4                                                 79713           105988          +32.96%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=1-4                102530          77457           -24.45%
BenchmarkRangeQuery/expr=-a_hundred,steps=100-4                                                                       3146151         2394632         -23.89%
BenchmarkRangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=1-4                                      208993          271831          +30.07%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-4                                                                102352034       78918095        -22.90%
BenchmarkRangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=10-4                                            49975           64390           +28.84%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-4                                                     172301          135386          -21.42%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-4                                                                     410919          522420          +27.13%
BenchmarkRangeQuery/expr=rate(a_ten[1m])_+_rate(b_ten[1m]),steps=10-4                                                 630188          497129          -21.11%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-4                                                                  11496338        9078827         -21.03%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=10-4                                              2368110         2978698         +25.78%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-4                                                                      67593           54078           -19.99%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-4                                                            196725783       245259240       +24.67%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-4                                                                   8123311         9973828         +22.78%
BenchmarkRangeQuery/expr=-a_one,steps=100-4                                                                           51339           41937           -18.31%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-4                                                                     74011           60475           -18.29%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1-4                                                  354956          428211          +20.64%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-4                                                             78094923        94113793        +20.51%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-4                                                    485114          402989          -16.93%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-4                                                            14770355        12314704        -16.63%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-4                                                                  176204546       147938590       -16.04%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-4                                                                   281513          236480          -16.00%
BenchmarkRangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=1-4                                          4026392         4754414         +18.08%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=100-4                                             8349225         9835313         +17.80%
BenchmarkRangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=1-4                                             43807           51199           +16.87%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-4                                                               2182909         1868284         -14.41%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=10-4               122017          104603          -14.27%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-4                                                              1300054518      1516171265      +16.62%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-4                                                          13176635974     15317115738     +16.24%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_ten[5m])),steps=1-4                                           1967549         2261152         +14.92%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-4                                                     401201          460892          +14.88%
BenchmarkRangeQuery/expr=label_join(a_ten,_'l2',_'-',_'l',_'l'),steps=1-4                                             225174          196189          -12.87%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=100-4                                                800644          698523          -12.75%
BenchmarkRangeQuery/expr=abs(a_ten),steps=100-4                                                                       495476          567173          +14.47%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-4                                                                 142490063       162405310       +13.98%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-4                                                           1405426830      1600650531      +13.89%
BenchmarkRangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=1000-4                                          876335          770699          -12.05%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1000-4                                               568898          645408          +13.45%
BenchmarkRangeQuery/expr=abs(a_ten),steps=10-4                                                                        222887          252545          +13.31%
BenchmarkRangeQuery/expr=-a_one,steps=1-4                                                                             32265           36517           +13.18%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-4                                                           99448527        88161270        -11.35%
BenchmarkRangeQuery/expr=rate(a_ten[1m])_+_rate(b_ten[1m]),steps=1000-4                                               10682731        12033128        +12.64%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=1000-4             14287756        16021103        +12.13%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-4                                                             29260262        26128723        -10.70%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-4                                                                 310938          277865          -10.64%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-4                                                              3801382         4243419         +11.63%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-4                                                                          1051529         942757          -10.34%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=10-4       4889650         5436450         +11.18%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_one[5m])),steps=10-4                                          352711          318496          -9.70%
BenchmarkRangeQuery/expr=-a_one,steps=1000-4                                                                          128053          115881          -9.51%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-4                                                               216374709       195917649       -9.45%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-4                                                              4779016         4327743         -9.44%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=1000-4                                    965622008       1062851298      +10.07%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-4     135185254       148767593       +10.05%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-4                                                          937024340       851534524       -9.12%
BenchmarkRangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=10-4                                         4771551         5249822         +10.02%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-4                                                             5135119         4704521         -8.39%
BenchmarkRangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=10-4                                        2130544         2314907         +8.65%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-4                                                                 2763631         2545205         -7.90%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-4                                             240331647       259565874       +8.00%
BenchmarkRangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=100-4                                        15635319        16881555        +7.97%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-4                                                                  14096411        15219593        +7.97%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-4                                                  9890678         9162296         -7.36%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-4                                                                   23677443        22033821        -6.94%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-4                                                               142181445       152705013       +7.40%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-4                                                              1481702323      1380144867      -6.85%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-4                                                                  438673          470740          +7.31%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-4                                                                19584025        20976292        +7.11%
BenchmarkRangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=1000-4                               45260467        48467454        +7.09%
BenchmarkRangeQuery/expr=label_join(a_ten,_'l2',_'-',_'l',_'l'),steps=10-4                                            250449          234060          -6.54%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-4                                                                  86494290        80865921        -6.51%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-4                                                  538572          504211          -6.38%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-4                                                   823800          878723          +6.67%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-4                                                  4554502         4840764         +6.29%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-4                                                                   952728          896457          -5.91%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-4                                                                    153509          144561          -5.83%
BenchmarkRangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=1000-4                                      46562876        49443686        +6.19%
BenchmarkRangeQuery/expr=-a_ten,steps=100-4                                                                           248249          263504          +6.15%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-4                                                                 7893151         8367415         +6.01%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-4                                                                    913128          967823          +5.99%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=100-4                                     111397261       118051970       +5.97%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_ten[5m])),steps=100-4                                         11009883        11662739        +5.93%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-4                                           6894876         7297041         +5.83%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=100-4              1801788         1905469         +5.75%
BenchmarkRangeQuery/expr=-a_one,steps=10-4                                                                            36880           38997           +5.74%
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-4                                                                      10033635        9501094         -5.31%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1-4        3812119         4021835         +5.50%
BenchmarkRangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=1000-4                                       122680502       116353156       -5.16%
BenchmarkRangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=10-4                                     237300          249600          +5.18%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-4                                            1895183888      1992827022      +5.15%
BenchmarkRangeQuery/expr=abs(a_ten),steps=1-4                                                                         181168          190412          +5.10%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=10-4                                                 442062          420671          -4.84%
BenchmarkRangeQuery/expr=abs(a_hundred),steps=100-4                                                                   4819000         5060188         +5.00%
BenchmarkRangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=1000-4                                   628201          659508          +4.98%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-4                                                    142026          149094          +4.98%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-4                                              2753333         2889056         +4.93%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-4                                               192059812       183176134       -4.63%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1000-4                                               3925819         3744912         -4.61%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-4                                             3325669         3175949         -4.50%
BenchmarkRangeQuery/expr=rate(a_ten[1m])_+_rate(b_ten[1m]),steps=100-4                                                1560827         1491086         -4.47%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-4                                                              136749670       142736575       +4.38%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-4                                                   603441          629466          +4.31%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-4                                                 2976505         2854762         -4.09%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_one[5m])),steps=100-4                                         1213673         1164104         -4.08%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-4                                                                1718995         1791463         +4.22%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-4                                                     153972          147941          -3.92%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_ten[5m])),steps=10-4                                          2694790         2804039         +4.05%
BenchmarkRangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=1-4                                                  77816           80948           +4.02%
BenchmarkRangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=1-4                                      52200           54296           +4.02%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-4                                            8028029         7732924         -3.68%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_ten[1m])),steps=1000-4                                                7000697         6744438         -3.66%
BenchmarkRangeQuery/expr=-a_ten,steps=10-4                                                                            177469          184128          +3.75%
BenchmarkRangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=100-4                                                205655          213332          +3.73%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-4                                                    189563          196472          +3.64%
…